% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{alicut}
\alias{alicut}
\title{wrapper to run Alicut}
\usage{
alicut(dir, ...)
}
\arguments{
\item{dir}{string. the directory your input files lie. working directory will
be temporarily set to it}

\item{...}{other parameters passed on to \code{\link[=system2]{system2()}}, including \code{args} and so
on.}
}
\value{
result of \code{\link[=system2]{system2()}}
}
\description{
wrapper to run Alicut
}
\examples{
\dontrun{
    aliscore('-i data-raw/aliscore/test.fasta')
    alicut('data-raw/aliscore', args = '-s')
}
}
