% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{aliscore}
\alias{aliscore}
\title{wrapper to run Aliscore}
\usage{
aliscore(args, ...)
}
\arguments{
\item{args}{character. options for \code{Aliscore.0x.y.pl}}

\item{...}{other parameters passed on to \code{\link[=system2]{system2()}}}
}
\value{
result of \code{\link[=system2]{system2()}}
}
\description{
wrapper to run Aliscore
}
\details{
aliscore will produce several files under the directory containg input files
}
\examples{
\\donotrun {
    aliscore('-i data-raw/aliscore/test.fasta')
}
}
