% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsxfun.R
\name{arr_slice.list}
\alias{arr_slice.list}
\title{A complicated version that slices along multiple axes}
\usage{
arr_slice.list(A, idx, axis = 1)
}
\description{
A complicated version that slices along multiple axes
}
