% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsxfun.R
\name{array2array}
\alias{array2array}
\title{Fix some weird data.frame}
\source{
https://stackoverflow.com/questions/49162079/why-does-array-returns-a-data-frame-as-it-is-in-r
}
\usage{
array2array(arr, callback = identical)
}
\description{
Fix some weird data.frame
}
