% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsxfun.R
\name{broadcast}
\alias{broadcast}
\title{Broadcasting an array to a wanted shape}
\usage{
broadcast(arr, dims)
}
\arguments{
\item{arr}{Array to be broadcasted}

\item{dims}{Desired output dimensions}
}
\value{
array of dimension dim
}
\description{
Extending a given array to match a wanted shape, similar to repmat.
}
\note{
# Fri Feb  9 14:41:25 2018 ------------------------------
}
\examples{
  M = array(1:12,dim=c(1,2,3))
  N = broadcast(M,c(4,2,3))

}
\author{
Feng Geng (shouldsee.gem@gmail.com)
}
