% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsxfun.R
\name{bsxfun}
\alias{bsxfun}
\title{Broadcasting an operation on multi-dimensinoal arrays}
\usage{
bsxfun(arrA, arrB, FUN = "+", ...)
}
\arguments{
\item{arrA, arrB}{Arrays to perfom the operator FUN(arrA,arrB)}

\item{FUN}{function to be performed (must be vectorised)}
}
\value{
An array of the shape with each dimension being the maximum of dimA,dimB.
}
\description{
Broadcasting two arrays to the same shape and then apply FUN.
}
\note{
# Fri Feb  9 14:41:25 2018 ------------------------------
}
\examples{

  M = array(1:12,dim=c(2,2,3))
  N = rbind(c(1,2))
  O = bsxfun(M,N,'*')
}
\author{
Feng Geng (shouldsee.gem@gmail.com)
}
