\name{calcCurve}
\alias{calcCurve}
\title{Calculate ROC curves based on ordered class data...}
\usage{calcCurve(data)
}
\description{Calculate ROC curves based on ordered class data}
\details{Calculate ROC curves based on ordered class data}
\value{Returns a data frame like data, but with additional columns TP and TN}
\author{Raivo Kolde <rkolde@gmail.com>}
\arguments{\item{data}{Data as described in \code{\link{data}}}
}

