\name{generate_gcocoa_query}
\alias{generate_gcocoa_query}
\title{Generate query for g:Cocoa web tool...}
\usage{generate_gcocoa_query(glist)
}
\description{Generate query for g:Cocoa web tool}
\details{Function that prints out a g:Gogoa query string given a list of gene name vectors.}
\value{Prints out the query string}
\author{Raivo Kolde <rkolde@gmail.com>}
\arguments{\item{glist}{a list of vectors of gene names}
}
\examples{glist = list(a = c("pax6", "klf9"), b = c("nanog", "Pou5f1"))}

