% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{methods}
\name{n_uniq}
\alias{n_uniq}
\title{Compute the number of unique elements}
\usage{
n_uniq(obj)
}
\arguments{
\item{obj}{object to compute distinct elements}
}
\value{
the number of uniq elements
}
\description{
Compute the number of unique elements
}
\examples{
 n_uniq(c(2, 3 ,2))
}
