% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ndchoose}
\alias{ndchoose}
\title{Multinomial coefficient}
\usage{
ndchoose(N, log = F)
}
\arguments{
\item{N}{a vector specifies the choice. For example N=c(5,1) specifies choose(6,1)}
}
\value{
The scalar multinomial coefficient
}
\description{
Multinomial coefficient
}
