% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\docType{methods}
\name{.paste}
\alias{.paste}
\alias{.paste,character,character-method}
\alias{\%+\%}
\title{Do some paste operation.}
\usage{
.paste(lhs, rhs)

\S4method{.paste}{character,character}(lhs, rhs)

lhs \%+\% rhs
}
\arguments{
\item{lhs}{the left hand side}

\item{rhs}{the right hand side}
}
\value{
returns a pasted value
}
\examples{
 "a" \%+\% "b"
}
