% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{saveplot}
\alias{saveplot}
\title{Save a plot to multiple places with multiple formats}
\usage{
saveplot(p, filename, out.folders, width = 10, height = 10,
  dpi = 720, format = c("svg", "eps", "png"))
}
\arguments{
\item{p}{the plot to save}

\item{filename}{name of the plot without suffix}

\item{out.folders}{vector of folders to plot}

\item{width}{width of plot in inch}

\item{height}{height of plot in inch}

\item{dpi}{dpi}

\item{format}{vector of file suffixes}
}
\description{
Takes a plot and saves it in multiple output folders using
 different file types.
}
