% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_color_continuous}
\alias{scale_color_continuous}
\title{Color scale}
\usage{
scale_color_continuous(set = "blue")
}
\arguments{
\item{set}{the color set to use}
}
\description{
Several color scalings.
}
