\name{theme_bw_raivo}
\alias{theme_bw_raivo}
\title{Theme_bw without ugly whitespace on x axis...}
\usage{theme_bw_raivo(base_size=12, base_family="")
}
\description{Theme_bw without ugly whitespace on x axis}
\details{Theme_bw without ugly whitespace on x axis}
\author{Raivo Kolde <rkolde@gmail.com>}
\arguments{\item{base_size}{base font size}
\item{base_family}{base font family}
}

