% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_general}
\alias{theme_general}
\title{Theme adopted from Jana}
\usage{
theme_general(title.hjust = 0, legend_pos = "bottom")
}
\arguments{
\item{title.hjust}{hjust of title}

\item{legend_pos}{position of legend}
}
\description{
This theme changes font sizes for plots used in papers
}
