% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unlist.df}
\alias{unlist.df}
\title{Unlist the "list" columns of a dataframe
Unlist the "list" columns of a dataframe, often resulted from as.data.frame(x) if x is a string array}
\usage{
unlist.df(df)
}
\description{
Unlist the "list" columns of a dataframe
Unlist the "list" columns of a dataframe, often resulted from as.data.frame(x) if x is a string array
}
