% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Write sequences to a fasta file.}
\usage{
write_fasta(fasta, path, width)
}
\arguments{
\item{fasta}{tibble. see \strong{Value} of \code{\link[=read_fasta]{read_fasta()}}.}

\item{path}{string. Path to the fasta file which you want write to.}

\item{width}{logical scalar. NotYetUsed.}
}
\value{
the input invisibly
}
\description{
overwrite
}
\examples{
{
    input_file  <- system.file('extdata', 'example.fasta', package = 'biozhuoer')
    output_file <- tempfile();
    write_fasta(read_fasta(input_file), output_file)
}

}
