% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_flex_read.R
\docType{package}
\name{xlsx_flex_read}
\alias{xlsx_flex_read}
\title{XLSX Flex Read}
\usage{
xlsx_flex_read(path, sheet = NULL, range = NULL, col_names = TRUE, ...)
}
\arguments{
\item{path}{path to the file to read as character vector.}

\item{sheet}{the workbook sheet as an index number or name of the sheet.}

\item{range}{A cell range to read. Only implemented for readxl::read_xlsx().}

\item{col_names}{Binary indicator to use thte first row as column names.}
}
\value{
A data.frame or tibble
}
\description{
This aims to provide a common interface for all three main XLSX reading functions. It defaults to this order, choosing the first available:
1. readxl::read_xlsx()
2. openxlsx::read.xlsx()
3. xlsx::read.xlsx()
}
\details{
XLSX Flex Read

Author: Daniel R. Williams
Date: 16 Feb 2024
}
\keyword{read}
\keyword{xlsx}
